// frontend/src/hooks/useProjectDetail.d.ts
import { Project } from "./useProjects";
import { Task } from "./useTasks";
import { Payment } from "./usePayments";
export interface ProjectDetail {
    project: Project;
    tasks: Task[];
    payments: Payment[];
    metrics: {
        total_tasks: number;
        completed_tasks: number;
        budget: number;
        payments: {
            facturado: number;
            pagado: number;
            saldo: number;
            egresos: number;
        };
    };
    important_dates: {
        start_date?: string | null;
        end_date?: string | null;
        next_task_start?: string | null;
        next_task_due?: string | null;
    };
}
export declare const useProjectDetail: (projectId?: string) => import("@tanstack/react-query").UseQueryResult<ProjectDetail, Error>;